pyproject.toml
[project]
name = "agent"
version = "0.0.2"
description = "Langgraph SQL Workflow"
authors = [
    { name = "Simon Cerda" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = [
    "langgraph==0.3.18",
    "python-dotenv>=1.0.1",
    "langchain-openai>=0.3.10",
    "psycopg2-binary>=2.9.10",
    "langchain>=0.3.21",
    "streamlit>=1.44.1",
    "sentence-transformers",
    "faiss-cpu",
    "sqlparse",
    "langgraph-cli[inmem]",
    "langchain-ollama",
    "sqlalchemy",
    
]


[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["langgraph.templates.agent", "agent"]
[tool.setuptools.package-dir]
"langgraph.templates.agent" = "src/agent"
"agent" = "src/agent"


[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"
